+++ packages/util/src/detectPackage.spec.ts
@@ -2,15 +2,13 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { assertSingletonPackage } from '.';
+import { detectPackage } from '.';
 
 describe('assertSingletonPackage', (): void => {
-  const NAME = 'assertSingletonPackage';
-
   it('should not log the first time', (): void => {
     const spy = jest.spyOn(console, 'warn');
 
-    assertSingletonPackage(NAME);
+    detectPackage({ name: '@polkadot/util', version: 'a.b.c-1st' });
     expect(spy).not.toHaveBeenCalled();
     spy.mockRestore();
   });
@@ -18,8 +16,8 @@ describe('assertSingletonPackage', (): void => {
   it('should log the second time', (): void => {
     const spy = jest.spyOn(console, 'warn');
 
-    assertSingletonPackage(NAME);
-    expect(spy).toHaveBeenCalledWith('Multiple versions of assertSingletonPackage detected, ensure that there is only one version in your dependency tree');
+    detectPackage({ name: '@polkadot/util', version: 'a.b.c-2nd' });
+    expect(spy).toHaveBeenCalledWith('Multiple versions of @polkadot/util detected: a.b.c-1st, a.b.c-2nd. Ensure that there is only one version in your dependency tree');
     spy.mockRestore();
   });
 });