+++ packages/util/src/index.ts
@@ -9,7 +9,7 @@ import './polyfill';
  */
 
 export { default as assert } from './assert';
-export { default as assertSingletonPackage } from './assertSingletonPackage';
+export { default as detectPackage } from './detectPackage';
 export { default as logger } from './logger';
 export { default as promisify } from './promisify';